Rename historicalState/index.ts to historicalState/HistoricalStateRegen.ts
